Oops, reverting one bad change from TS-620
git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@1056247 13f79535-47bb-0310-9956-ffa450edef68

+++ proxy/InkAPI.cc
@@ -1752,7 +1752,7 @@ _TSrealloc(void *ptr, size_t size, const char *path)
 
 // length has to be int64_t and not size_t, since -1 means to call strlen() to get length
 char *
-_TSstrdup(const char *str, int length, const char *path)
+_TSstrdup(const char *str, int64_t length, const char *path)
 {
   return _xstrdup(str, length, path);
 }
